Rename GDPR to Privacy, consolidate functions
