+++ hooks.php
@@ -294,5 +294,5 @@ Container::get( 'Styles' )->init();
 // GDPR
 // -------------------------------------------------------------------------------------------------------------------
 
-add_action( 'init', [ '\Pressbooks\Gdpr', 'init' ], 9 ); // Must come before `add_action( 'init', 'wp_schedule_delete_old_privacy_export_files' );`
+add_action( 'init', [ '\Pressbooks\Privacy', 'init' ], 9 ); // Must come before `add_action( 'init', 'wp_schedule_delete_old_privacy_export_files' );`
