+++ requires-admin.php
@@ -13,4 +13,3 @@ require_once( __DIR__ . '/inc/admin/metaboxes/namespace.php' );
 require_once( __DIR__ . '/inc/admin/networkmanagers/namespace.php' );
 require_once( __DIR__ . '/inc/admin/organize/namespace.php' );
 require_once( __DIR__ . '/inc/admin/plugins/namespace.php' );
-require_once( __DIR__ . '/inc/admin/privacy/namespace.php' );
